Updating a number of dependencies Fixes some issues with @IntDefs using long in some cases, also fixes a couple of method signatures to properly accept integers Fixes some nullability issues after upgrading to the latest support library

+++ app/src/main/kotlin/com/kelsos/mbrc/content/library/albums/AlbumRepository.kt
@@ -11,7 +11,7 @@ interface AlbumRepository : Repository<AlbumEntity> {
    * Retrieves the albums ordered by
    */
   fun getAlbumsSorted(
-      @Sorting.Fields order: Long = Sorting.ALBUM_ARTIST__ALBUM,
-      ascending: Boolean = true
+    @Sorting.Fields order: Int = Sorting.ALBUM_ARTIST__ALBUM,
+    ascending: Boolean = true
   ): Single<DataSource.Factory<Int, AlbumEntity>>
 }
